requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
